Fix the review-2026-09 findings: proven branch-isolation race, replay-of-injection semantics, admin token narrowing, and the hygiene pass.
P0 — branch isolation (new invariant 8): fan-out siblings share the underlying msg.Decoded/msg.Meta maps, and the Starlark remove() glue's lazy path deleted from that shared Go map in place — a sibling branch encoding the same map hit the unrecoverable fatal error: concurrent map iteration and map write (PoC: source fanned out to a remove-script transform and a json sink; first batch killed the process). deleteKey now materializes the copy-on-write tree before deleting, like every other mutation path; the message-ownership contract (transforms replace, never mutate in place) is documented on registry.Message and pkg/plugin; TestInvariant_BranchIsolation locks it in.

P1 — engine correctness: crash-replay of an internal injection (testkit capture, DLQ reinjection) was dispatched as a fan-out from its node — skipping the transform or counting as filtered at a sink; replay now re-enters the node through the shared dispatchInternal path, honoring meta.injected_at (replay_inject_test.go). writeBatch takes the max explicit timeout across a mixed batch instead of letting the last edge win. engine.New normalizes WasmSlowCallWarnMs (0 → 5000ms, negative disables) like every other option.

P2 — security and hygiene: the ?token= query form is accepted on /admin/sse only and every response carries X-Content-Type-Options: nosniff; Engine.Run is single-shot (CAS) and Abandon returns (int, error) with the job runner propagating the failure; transform channels size buffers with the same surge-cap rule as sinks; DeadLetter drops the never-populated origin_node/retry_count; store.NewMemory() loses its misleading pipeline argument; .gitattributes (text=auto eol=lf + binary marks) fixes golden-file tests on autocrlf checkouts; the hand-rolled insertion sort in internal/ir, the admin placeholder statement and the unused workers computation are gone; the CLI help header drops "v3 POC".

CHANGELOG: Unreleased section records the pass. Full suite green (all packages), engine -race clean, golangci-lint 0 issues.

+++ internal/lang/starhost/remove_test.go
@@ -5,6 +5,9 @@ import "testing"
 // remove() is the del() migration glue (redesign-v3.md §4.8): it must work
 // on the lazy root (marking the state dirty so the engine writes the
 // deletion back), on materialized trees, and be a no-op for missing keys.
+// Even a lazy-root remove materializes the COW tree first (review-2026-09:
+// in-place delete would race sibling fan-out branches), so values roundtrip
+// to JSON-ish types (int64) exactly like any SetKey write-back.
 func TestRemoveOnLazyRoot(t *testing.T) {
 	ps, _, serr := runScript(t, `remove(payload, "temp")`,
 		map[string]any{"temp": "x", "keep": 1}, map[string]any{})
@@ -18,7 +21,7 @@ func TestRemoveOnLazyRoot(t *testing.T) {
 	if _, exists := got["temp"]; exists {
 		t.Fatalf("temp not removed: %v", got)
 	}
-	if got["keep"] != 1 {
+	if got["keep"] != int64(1) { // int64: lazy remove roundtrips via the COW tree
 		t.Fatalf("keep lost: %v", got)
 	}
 }
